Add getTiltDegrees call
There is a function missing to retrieve the current angle of kinect's
motor. This commit add it to the haskell interface.

For this we introduce a new datatype for freenect_raw_tilt_state*
pointer and use the given libfreenect interface to get it from the
corresponding instance.

It also automatically updates the internal tilt state, where the
error code is not catched. (libfreenect's examples are also not catching
this failure in general).